changes in pagination data type to long

+++ src/main/java/com/cctns/autosave/producer/service/core/domain/PageDomain.java
@@ -12,5 +12,5 @@ import lombok.ToString;
 public class PageDomain<T>{
     T list;
     Long totalSize;
-    Integer pageCount;
+    Long pageCount;
 }
